pyproject.toml
[project]
name = "freelance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.37.1",
    "jsonpatch>=1.33",
    "langchain-core>=0.3.83",
    "langgraph>=0.2.35",
    "loguru>=0.7.2",
    "openpyxl>=3.1.5",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
]
